+++ History.md
@@ -1,3 +1,6 @@
+===================
+  * removed using namespace v8 (bad practice)
+
 0.0.8 - Oct 14 2013
 ===================
   * Epoll thread code improved (#4)